abi: introduce TaskTableIndex
TaskTableIndex is a lightweight wrapper around a RawTaskTableIndex
(currently a u16) that restricts allowed values to those representable
in a 10-bit unsigned integer (bit width of TaskId's index component).

TaskIds are composed of an index and generation packed into a RawTaskId
(currently a u16) which uniquely identify a running instance of a task.
Occasionally, there is need to refer to a task without referring to a
specific running instance.  Since the index component of a TaskId is
simply an integer index into the kernel's app task table,
TaskId::index() provides the index component of a TaskId as a usize for
this purpose and TaskId::for_index_and_gen() takes a usize argument when
reconstructing a TaskId from its components.  This presents two
problems: usize can store values outside the range allowed by the packed
TaskId format and its difficult to tell when a usize is actually a task
index and should be kept within the allowable range.  TaskTableIndex
provides a type to replace this use of usize.